move ts-config test to ts folder, follow class variable name cammel case
