+++ test/typescript/typescript-config.test.ts
@@ -2,12 +2,12 @@ import { writeFileSync, mkdirSync } from "fs";
 import * as path from "path";
 import * as semver from "semver";
 import * as ts from "typescript";
+import { NodeProject } from "../../src/javascript";
 import {
-  NodeProject,
   TypescriptConfig,
   TypescriptConfigExtends,
   TypeScriptModuleResolution,
-} from "../../src/javascript";
+} from "../../src/typescript";
 import { withProjectDir } from "../util";
 
 describe("TypescriptConfig", () => {
